# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RAGO API
  description: |
    RAGO is a local-first, privacy-first Go AI foundation with four core pillars:
    - **LLM**: Language model management and generation
    - **RAG**: Retrieval-augmented generation
    - **MCP**: Model Context Protocol for tool integration
    - **Agents**: Autonomous workflow orchestration
  version: 3.0.0
  contact:
    name: RAGO Support
    url: https://github.com/liliang-cn/rago
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:7127/api/v1
    description: Local development server
  - url: https://api.rago.ai/v1
    description: Production server
...
paths:
...
  /rag/ingest:
    post:
      tags: [RAG]
      summary: Ingest document
      description: Ingest a document into the RAG system
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Document ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /rag/search:
    post:
      tags: [RAG]
      summary: Search documents
      description: Search for relevant documents using vector similarity
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /rag/documents:
    get:
      tags: [RAG]
      summary: List documents
      description: Get a list of all documents in the RAG system
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: source
          in: query
          schema:
            type: string
          description: Filter by source
        - name: type
          in: query
          schema:
            type: string
          description: Filter by document type
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Limit number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: Documents listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # MCP endpoints
  /mcp/tools:
    get:
      tags: [MCP]
      summary: List tools
      description: Get a list of all available MCP tools
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Tools listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /mcp/tools/{name}/call:
    post:
      tags: [MCP]
      summary: Call a tool
      description: Execute an MCP tool with the provided parameters
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Tool name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolCallRequest'
      responses:
        '200':
          description: Tool executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolCallResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # Agent endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    IngestRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        source:
          type: string
        metadata:
          type: object
    IngestResponse:
      type: object
      properties:
        document_id:
          type: string
        chunks_created:
          type: integer
        status:
          type: string
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 10
        threshold:
          type: number
          default: 0.7
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              content:
                type: string
              score:
                type: number
              metadata:
                type: object
    Document:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        content:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
    ToolInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
    ToolCallRequest:
      type: object
      required: [parameters]
      properties:
        parameters:
          type: object
    ToolCallResponse:
      type: object
      properties:
        result:
          type: object
        error:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string